requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.8.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
flake8-black==0.3.3
idna==3.4
kombu==5.2.4
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
paypalrestsdk==1.13.1
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.33
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
